v 0.1

+++ UserAvatar.Bll.TaskManager/Services/CommentService.cs
@@ -29,53 +29,112 @@ namespace UserAvatar.Bll.TaskManager.Services
             _mapper = mapper;
         }
 
-        public async Task<CommentModel> CreateNewCommentAsync(int userId, int cardId, string text)
+        public async Task<Result<CommentModel>> CreateNewCommentAsync(
+            int userId, int boardId, int cardId, string text)
         {
-            await ValidateUserByCardAsync(userId, cardId);
+            if (!await _boardStorage.IsBoardExistAsync(boardId))
+            {
+                return new Result<CommentModel>(ResultCode.NotFound);
+            }
+
+            if (!await _boardStorage.IsUserBoardAsync(userId, boardId))
+            {
+                return new Result<CommentModel>(ResultCode.Forbidden);
+            }
+
+            if (!await _boardStorage.IsBoardCard(boardId, cardId))
+            {
+                return new Result<CommentModel>(ResultCode.Forbidden);
+            }
+            //await ValidateUserByCardAsync(userId, cardId);
 
             var newComment = new Comment
             {
                 CardId = cardId,
                 UserId = userId,
                 Text = text,
-                CreatedAt = DateTime.Now,
-                ModifiedAt = DateTime.Now
+                CreatedAt = DateTimeOffset.Now,
+                ModifiedAt = DateTimeOffset.Now
             };
+            // TODO: It returns the comment, but shouldn't , let's return nothing
+            var comment = await _commentStorage.CreateAsync(newComment);
 
-            await _commentStorage.CreateAsync(newComment);
-
-            return _mapper.Map<Comment, CommentModel>(newComment);
+            return new Result<CommentModel>(_mapper.Map<Comment, CommentModel>(comment));
         }
 
-        public async Task<CommentModel> UpdateCommentAsync(int userId, int commentId, string text)
+        public async Task<Result<CommentModel>> UpdateCommentAsync(
+            int userId, int boardId, int cardId, int commentId, string text)
         {
-            await ValidateUserByCommentAsync(userId, commentId);
+            if (!await _boardStorage.IsBoardExistAsync(boardId))
+            {
+                return new Result<CommentModel>(ResultCode.NotFound);
+            }
+
+            if (!await _boardStorage.IsUserBoardAsync(userId, boardId))
+            {
+                return new Result<CommentModel>(ResultCode.Forbidden);
+            }
+
+            if (!await _boardStorage.IsBoardCard(boardId, cardId))
+            {
+                return new Result<CommentModel>(ResultCode.Forbidden);
+            }
+
+            if (!await _cardStorage.IsCardComment(cardId, commentId))
+            {
+                return new Result<CommentModel>(ResultCode.Forbidden);
+            }
+            
+            //await ValidateUserByCommentAsync(userId, commentId);
 
             var thisComment = await _commentStorage.GetCommentByCommentIdAsync(commentId);
             thisComment.Text = text;
 
             await _commentStorage.UpdateCommentAsync(thisComment);
 
-            return _mapper.Map<Comment, CommentModel>(thisComment);
+            return new Result<CommentModel>(_mapper.Map<Comment, CommentModel>(thisComment));
         }
 
-        public async Task<Result<List<CommentModel>>> GetCommentsAsync(int userId, int cardId)
+        /*public async Task<Result<List<CommentModel>>> GetCommentsAsync(int userId, int cardId)
         {
             await ValidateUserByCardAsync(userId, cardId);
 
             var commentList = await _commentStorage.GetAllAsync(cardId);
 
             return new Result<List<CommentModel>>(_mapper.Map<List<Comment>, List<CommentModel>>(commentList));
-        }
+        }*/
 
-        public async Task DeleteCommentAsync(int userId, int commentId)
+        public async Task<int> DeleteCommentAsync(
+            int userId, int boardId, int cardId, int commentId)
         {
-            await ValidateUserByCommentAsync(userId, commentId);
+            if (!await _boardStorage.IsBoardExistAsync(boardId))
+            {
+                return ResultCode.NotFound;
+            }
+
+            if (!await _boardStorage.IsUserBoardAsync(userId, boardId))
+            {
+                return ResultCode.Forbidden;
+            }
+
+            if (!await _boardStorage.IsBoardCard(boardId, cardId))
+            {
+                return ResultCode.Forbidden;
+            }
+
+            if (!await _cardStorage.IsCardComment(cardId, commentId))
+            {
+                return ResultCode.Forbidden;
+            }
+            
+            //await ValidateUserByCommentAsync(userId, commentId);
             await _commentStorage.DeleteApparentAsync(commentId);
+            return ResultCode.Success;
         }
 
         private async Task ValidateUserByCommentAsync(int userId, int commentId)
         {
+            
             var cardId = await _commentStorage.GetTaskIdByCommentIdAsync(commentId);
             var isUserInThisBoard = await _boardStorage.IsUserBoardAsync(userId, await _cardStorage.GetBoardIdAsync(cardId));
             if (!isUserInThisBoard)
